rename: canvas -> draw
