+++ packages/repl/package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "@strudel/codemirror": "workspace:*",
     "@strudel/core": "workspace:*",
-    "@strudel/canvas": "workspace:*",
+    "@strudel/draw": "workspace:*",
     "@strudel/hydra": "workspace:*",
     "@strudel/midi": "workspace:*",
     "@strudel/mini": "workspace:*",
